+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * __DebugAdapter.raise_event(event,data,modname) to call event handlers directly for testing
 * Updated LuaObject recongition for debug views to reflect implementation detail changes in Factorio's API (no longer uses `__self`)
 * Don't accept `nil` filters on custom-input events
+* TypeDef generator now accepts Factorio Machine-Readable-Docs V2
+* Various disassembler view fixes
+
 ## 1.1.11
 
 * Don't offer ! and version in depencency snippets