+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 1.1.13
 
+* Added `__REMARK_COLOR_BEGIN__` and `__REMARK_COLOR_END__` to locale highlighter
 ## 1.1.12
 
 * __DebugAdapter.terminate() to end session from script