+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [<img height='36' style='border:0px;height:36px;' src='https://az743702.vo.msecnd.net/cdn/kofi2.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com'/>](https://ko-fi.com/X8X41IE4T)
 
+## 1.1.15
+
+
 ## 1.1.14
 
 * Load LuaObject type data from Factorio's JSON docs for listing properties when debugging